refactor(command-visualizer): rewire logic/ imports to model/
Repoint logic/ (pipeline, fixture-extract, hat-allocator, tokenize, chain
+ their tests) at ../model/* and ../data/* for the shared contract and
constants. Pos/Range now come from ../model/geometry. No logic/ file
imports render/.

Pure import rewrite; no behavior change.
